+++ src/ui/NoData/NoData.tsx
@@ -2,11 +2,15 @@ import Typography from '@mui/material/Typography'
 import DraftsTwoToneIcon from '@mui/icons-material/DraftsTwoTone'
 import { Stack } from '@mui/joy'
 
-export default function NoData(): JSX.Element {
+interface NoDataProps {
+  colSpan?: number
+}
+
+export default function NoData({ colSpan = 6 }: NoDataProps): JSX.Element {
   return (
     <tr>
-      <td colSpan={6} style={{ textAlign: 'center', padding: '20px 0' }}>
-        <Stack display='row' alignItems='center'>
+      <td colSpan={colSpan} style={{ textAlign: 'center', padding: '20px 0' }}>
+        <Stack direction='row' alignItems='center' justifyContent='center' spacing={1}>
           <DraftsTwoToneIcon color='primary' sx={{ fontSize: 40 }} />
           <Typography variant='overline' color='textSecondary'>
             No Data